// frontend/src/app/infraestructure/models/presupuestos/filters/m_calendar.ts
import { z } from 'zod';
const meses = z.object({
  mes: z.string(),
  value: z.string(),
});

export const ItemsRCalendar = z.object({
  meses: z.array(meses),
  mesActual: z.string(),
  asigado: z.boolean(),
  TotalAsignado: z.number(),
});

export type DTO_Pre_CalendarR = z.infer<typeof ItemsRCalendar>;